base form login

+++ src/views/pages/login/Login.js
@@ -1,13 +1,11 @@
 import React, { useState, useEffect } from 'react'
 import { useNavigate } from 'react-router-dom'
-import { Link } from 'react-router-dom'
 import axios from 'axios'
 import sweetalert from 'sweetalert2'
 import {
   CButton,
   CCard,
   CCardBody,
-  CCardGroup,
   CCol,
   CContainer,
   CForm,
@@ -38,7 +36,16 @@ const Login = () => {
         username: username,
         password: password
       })
-      console.log('Login successful:', response.data)
+      console.log(response.data)
+      if (response.data === 'Invalid username or password') {
+        sweetalert.fire({
+          icon: "error",
+          title: "Sai tên đăng nhập hoặc mật khẩu",
+          showConfirmButton: false,
+          timer: 2000
+        })
+        return
+      }
       if (response.data.sessionId) {
         sessionStorage.setItem('sessionId', response.data.sessionId)
         let sessionId = sessionStorage.getItem('sessionId')
@@ -49,8 +56,7 @@ const Login = () => {
             showConfirmButton: false,
             timer: 2000
           }).then((result) => {
-            /* Read more about handling dismissals below */
-            if (result.dismiss === sweetalert.DismissReason.timer) {
+            if (result.dismiss === sweetalert.DismissReason.timer || result.dismiss === sweetalert.DismissReason.backdrop) {
               window.location.href = '/'
             }
           });
@@ -70,75 +76,51 @@ const Login = () => {
   }
 
   return (
-    <div className="bg-body-tertiary min-vh-100 d-flex flex-row align-items-center">
+    <div className="bg-body-tertiary min-vh-100 d-flex align-items-center justify-content-center">
       <CContainer>
         <CRow className="justify-content-center">
-          <CCol md={8}>
-            <CCardGroup>
-              <CCard className="p-4">
-                <CCardBody>
-                  <CForm>
-                    <h1>Login</h1>
-                    <p className="text-body-secondary">Sign In to your account</p>
-                    <CInputGroup className="mb-3">
-                      <CInputGroupText>
-                        <CIcon icon={cilUser} />
-                      </CInputGroupText>
-                      <CFormInput
-                        placeholder="Username"
-                        autoComplete="username"
-                        value={username}
-                        onChange={(e) => setUsername(e.target.value)}
-                      />
-                    </CInputGroup>
-                    <CInputGroup className="mb-4">
-                      <CInputGroupText>
-                        <CIcon icon={cilLockLocked} />
-                      </CInputGroupText>
-                      <CFormInput
-                        type="password"
-                        placeholder="Password"
-                        autoComplete="current-password"
-                        value={password}
-                        onChange={(e) => setPassword(e.target.value)}
-                      />
-                    </CInputGroup>
-                    <CRow>
-                      <CCol xs={6}>
-                        <CButton
-                          color="primary"
-                          className="px-4"
-                          onClick={handleLogin} // Gọi handleLogin khi nhấn nút
-                        >
-                          Login
-                        </CButton>
-                      </CCol>
-                      <CCol xs={6} className="text-right">
-                        <CButton color="link" className="px-0">
-                          Forgot password?
-                        </CButton>
-                      </CCol>
-                    </CRow>
-                  </CForm>
-                </CCardBody>
-              </CCard>
-              <CCard className="text-white bg-primary py-5" style={{ width: '44%' }}>
-                <CCardBody className="text-center">
-                  <div>
-                    <h2>Sign up</h2>
-                    <p>
-                      Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod
-                      tempor incididunt ut labore et dolore magna aliqua.
-                    </p>
-                    <Link to="/register">
-                      <CButton color="primary" className="mt-3" active tabIndex={-1}>
-                        Register Now!
+          <CCol md={6} lg={5} xl={4}>
+            <CCard className="p-4 shadow-lg border-0">
+              <CCardBody>
+                <CForm>
+                  <h1 className="text-center mb-4">Đăng Nhập</h1>
+                  <CInputGroup className="mb-3">
+                    <CInputGroupText>
+                      <CIcon icon={cilUser} />
+                    </CInputGroupText>
+                    <CFormInput
+                      placeholder="Tên đăng nhập"
+                      autoComplete="username"
+                      value={username}
+                      onChange={(e) => setUsername(e.target.value)}
+                    />
+                  </CInputGroup>
+                  <CInputGroup className="mb-4">
+                    <CInputGroupText>
+                      <CIcon icon={cilLockLocked} />
+                    </CInputGroupText>
+                    <CFormInput
+                      type="password"
+                      placeholder="Mật khẩu"
+                      autoComplete="current-password"
+                      value={password}
+                      onChange={(e) => setPassword(e.target.value)}
+                    />
+                  </CInputGroup>
+                  <CRow>
+                    <CCol xs={12} className="text-center">
+                      <CButton
+                        color="primary"
+                        className="px-5 py-2"
+                        onClick={handleLogin} // Gọi handleLogin khi nhấn nút
+                      >
+                        Đăng nhập
                       </CButton>
-                    </Link>
-                  </div>
-                </CCardBody>
-              </CCard>
-            </CCardGroup>
+                    </CCol>
+                  </CRow>
+                </CForm>
+              </CCardBody>
+            </CCard>
           </CCol>
         </CRow>
       </CContainer>
